nanoMIPS: Fix statx padding
There is some missing padding in struct statx_timestamp (__reserved is
required or uint32_t stx_rdev_major will land where __reserved would
be), and struct statx (__spare2[14] is required or the whole structure
will be the wrong size and writes to it by the statx system call will
overflow the buffer). For consistency we also add __spare0[1] which
isn't strictly required due to alignment of the uint64_t stx_ino after
it.

Signed-off-by: James Hogan <james.hogan@imgtec.com>