Use stat syscall wrappers where possible
There are several uses of fstatat, lstat, and fstat system calls using
__syscall which would get further complicated by newer architecture ABIs
which don't expose the older stat calls and expect them all to be
implemented using the newer statx system call.

To avoid the complexity, use the existing wrappers of these calls which
should already do the right thing depending on the available system
calls.

Signed-off-by: James Hogan <james.hogan@imgtec.com>