stat: move statx handling out of __syscall5
Drop the unpleasant hack in __syscall5 to conditionally convert the
struct statx produced by the statx system call to a struct stat, and do
this in the individual wrappers.

This is much more intuitive, less subtle, and may be important in future
if we ever need a statx wrapper. Most importantly however the wrappers
themselves are generic (they should really work out of the box on other
architectures) and __syscall5 is architecture specific.

Signed-off-by: James Hogan <james.hogan@mips.com>