fall back to statx syscall in stat wrappers
Fall back to using the statx system call if the older stat system calls are
not available, which would be the case for newer system call ABIs.

Also convert the struct statx produced by the statx system call to a
struct stat in the individual wrappers, if necessary.

This helps keep the wrappers generic, so that they will work out of the box on
newer architectures which only support statx. This is also much more intuitive,
less subtle, and may be important in the future if we need a statx wrapper.